+++ test/integration/test_call_novel_orf.py
@@ -17,6 +17,8 @@ def create_base_args() -> argparse.Namespace:
     args.annotation_gtf = None
     args.proteome_fasta = None
     args.reference_source = None
+    args.codon_table = 'Standard'
+    args.chr_codon_table = ['chrM:SGC1']
     args.output_path = None
     args.output_orf = None
     args.coding_novel_orf = False