+++ CHANGELOG.md
@@ -12,7 +12,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 - Added `--codon-table` and `--chr-codon-table`. The former sets the codon table to use, and the latter overrides it for specific chromosomes.
 
-- Added the support for codon table for callVariant.
+- Added the support for codon table for `callVariant`.
+
+- Added the support for codon table for `callNovelORF` and `callAltTranslation`.
 
 ## [1.4.6] - 2025-05-21